Language
The "not covered" threw me off a bit as both lowercase and uppercase are referenced(covered) above. I went off on a tangent looking for a bold and italics filters to chain in thinking those were the non-covered aspects from the gif sample. Also removed the ** as there is no footnote which it appeared to be denoting.